requirement.txt
langchain
langchain-openai
langchain-community
langchain-core
pypdf
chromadb
tiktoken
python-dotenv
aiohttp
sentence_transformers